랜딩 게임중 다음타자, 좌측메뉴 커뮤니티 업데이트

+++ Assets/Scripts/Common/Models/TeamInfo.cs
@@ -222,9 +222,9 @@ public class TeamInfo {
 		}
 	}
 
-	int _behind;
+	float _behind;
 
-	public int behind {
+	public float behind {
 		get {
 			return _behind;
 		}
